[Split shade] Fix QS actions footer border visible on shade expansion
* Also fixes the actions footer fading in too late on split shade

The QS actions footer has its own background and elevation, which makes
it become visible during the split shade expansion.

The fix is to make sure that there different expansion values are used
for the actions footer when on split shade, so that the background of
the actions footer fades in later.

Change-Id: I81dcba67b9473d9a8a4565dcdffd0a62f3817a50
Test: FooterActionsControllerTest.kt
Test: QSFragmentTest.java
Test: Manually
Fixes: 240563302