HeadsUpViewBinder: Don't WTF if notif stops alerting after cleanup.
Right now, a notification can go from heads up to not after it's removed
from the NotifCollection and cleaned up, and HeadsUpCoordinator will
happily ask HeadsUpViewVinder to unbind the heads up view, but then
RowContentBindStage will complain that there are no bind stage params
for the notification.

Long-term, we'd like to improve the clarity of the interactions between
a notification's lifetime in NotifCollection and in HeadsUpManager, but
short-term, this change keeps HeadsUpViewBinder from triggering a WTF
when a notification stops alerting after it was cleaned up.

Bug: 245537746
Test: atest HeadsUpViewBinderTest
Change-Id: I06a6a74c9946398cc4ed22544e5d205f6d74bd6a