Disable the Accessibility Shortcut dialog on TV.
When the user presses the accessibility shortcut (e.g. to enable the screen-reader), some dialog would show up that asked if the user really wanted to do that. This behaviour is confusing to users. Instead, on TV we want to directly trigger the shortcut action whenever the shortcut is performed.

Bug: b/238744704
Test: Locally tested on adt4-userdebug
Change-Id: I85567c4db7632ee51e8e29ce753ed7bda386f4e2