Clarify that config_defaultFirstUserRestrictions is for the SYSTEM user
Default user restrictions can be set by user type. For the SYSTEM user, i.e. user 0, this is done via setting config_defaultFirstUserRestrictions. The meaning of "FirstUser" is slightly ambiguous - on a headless-system-user it isn't totally clear what that means. So we update the documentation here to clarify that it is specifically for the SYSTEM user (user 0).

Bug: 251802540
Change-Id: Id0ec0153f9637aace192fd0e2c3e70fa1c206b92
Test: N/A