Remove TextAppearance.QS.Status values-night entry
This text style was configured in values/ to use
`config_headlineFontFamily` and in values-night/ to use
`config_bodyFontFamily`. It looks like maybe BatteryMeterView was the
only view that updated itself on that config value change, so this
resulted in the battery percent text changing its stroke width on night
mode change.

Test: manual
Fixes: 254107156
Change-Id: Ifcfe78162381c8a220078c8532b004a8ed492572