Provide startBounds for closing container with legacy transition
Before, we only provide startBounds when the animation target is
MODE_CHANGING, with which we won't know the start bounds for target that
is closing while resizing.

Now, we provide startBounds for all targets, so that we can use it to
provide seamless animation when the target is closing while resizing.

Also make sure the closing container's surface is in the start bounds
before the animation starts to avoid any flicker.

Bug: 241043533
Test: Verify with switching between split and stack.
Change-Id: I4d8e1ee44d418249dcb434425c7ab3d63a429bad