Add a y translation to clock complication.
The clock complication comes with internal paddings that make the
spacing with the complications below too large, even without any layout
engine margin. Without a better solution, add a y translation to the
clock so reduce the spacing, and also change its text size to dp unit
so it doesn't respond to system text size change, which would mess up
the translation.

Test: https://screenshot.googleplex.com/BLfQEhA6i2t4qyu
Test: https://screenshot.googleplex.com/BbuvDvfmahtNy5U
Bug: 257553229
Change-Id: I7350981a28af590f6e70fb4fab004d5180cda502