Fixes user switcher dialog animations.
The user switcher dialog uses a chain of dialogs linked through common
CUJs to morph-animate dialogs into one another as the user goes through
flows that require showing a second dialog after the first.

ag/20259287 was an overzealous fix for b/254704051; a bug where the user
switcher dialog was still visible above the "add supervised user"
activity. The fix addressed the problem but introduced a regression
because it was dismissing the user switcher dialog each time any item
in the dialog was clicked.

This CL basically reverts ag/20259287 and re-does the fix more
surgically in a lower layer. UserInteractor will ask to dismiss the
dialog if it wants to start the activity for the specific action of
adding a supervised user.

Fix: 262743114
Test: modified UserInteractorTest (and also modernized it all)
Test: manually verified that the dialogs morph-animate
Test: manually verified that starting the "add supervised user" activity
causes the user switcher dialog to be dismissed
Test: all tested on handheld device (phone)

Change-Id: Ia33c8644ca891879f36c62c0257ebdc897623c40