Device provisioned / user setup complete changes in HsumBootUserInitializer.
With this change, we no longer set the DEVICE_PROVISIONED flag in
HsumBootUserInitializer, as the flag is set at the end of setup,
as intended.

For consistency with the non-HSUM configuration, we also set
USER_SETUP_COMPLETE for user 0 only when the device has been set up
at least once (i.e. only when DEVICE_PROVISIONED is true).

Bug: 261791491
Test: flash, check flag values, set up device, check flags again
Change-Id: I7e11661864200c477168b1089d14b5476736e597