Fix NotificationTest on small screen.
The send button may be hidden by IME window if the screen size is small.
This CL dismisses IME to help UiAutomator find the send button.

An alternative approach would be to scroll up the notification.
However, the fact that IME window covers notification makes it hard to
scroll the notification programatically.

Bug: 265259897
Bug: 269715255
Test: atest InputMethodStressTest:NotificationTest#testDirectReply
Change-Id: Iad6dd98b01a3ba6824c1bc3ad8b4d9c9a648ad96