Create InputManagerGlobal class
InputManager should  always have a context associated with it instead
of injecting it sometimes by calling InputManager#getInstance(Context)
as was done as a part of b/266013036.
Uptil  now, InputManager was a singleton class and it contained the
IInputManager interface (which talks to the system server service).
This CL introduces InputManagerGlobal, which will be the singleton class
(with the IInputManager interface).

Bug: b/267758905
Test: Pre-submit
Change-Id: I1d24e2b5692d0eb7d4c23c9b6a8e1a325eaaad83