Skip merging non-ready transition
Otherwise TransitionHandler#mergeAnimation will receive
null transition info and null transaction. And if it is
passed to remote, the remote may meet NPE on binder thread
and never report transition finish. Which will lead to
a permanent animating state and affect activity life cycle,
e.g. unable to be stopped.

Bug: 269607576
Test: ShellTransitionTests#testTransitionMergingOnFinish
Change-Id: I391ce78c6dcb7ed7bfef41d33d1e633dba3d680a