AnrTest: Find target object by package
The test used to look for the action bar in order to dispatch a
MotionEvent for triggering ANR. Since Wear doesn't support action bars,
the test was failing for Wear devices. With this change, the test now
looks for any view in the test activity (i.e. look by package, instead
of the action bar).

Bug: 269148947
Test: atest AnrTest - passes on a Wear device
Change-Id: Ic9eb44db14ffe40c5553563e0f93fda4cefc9ac5