[Sb refactor] Add carrierId to the new pipeline
Adds carrierId to the mobile connection repo, and upgrades the mobile
icon group type to a NetworkTypeIconModel so that we can do fancy stuff
like override it.

We add support for checking MobileIconCarrierIdOverrides for an override
entry whenever the carrierId changes. If an override entry exists, then
we check for an asset override whenever the network changes.

Also can be tested using Demo mode:
```
$ # MCC/MNC 311/480 overrides the icon to use 5GUC -- no carrierId
$ # override for 1234
$ adb shell am broadcast -a com.android.systemui.demo -e command network -e mobile show -e level 2 -e datatype 5g+ -e mccmnc 311480 -e carrierid 1234
$ # carrierid 2032 contains an icon override, pointing back to the
$ # default 5G+ icon
$ adb shell am broadcast -a com.android.systemui.demo -e command network -e mobile show -e level 2 -e datatype 5g+ -e mccmnc 311480 -e carrierid 2032
```

Test: MobileConnectionsRepositoryTest
Test: MobileIconInteractorTest
Bug: 272099695
Change-Id: Ia71831d481d84c2508d8cbe08755644eed2f9ff5