Resolve requested deferral policy based on caller.
Since there's going to be a very long-tail of broadcasts from the OS
sent to apps that we should mark as "defer until active", support the
ability to let a sender force a given policy, otherwise fall back to
reasonable defaults.

To mitigate risk of this CL landing, we're merging this as disabled,
and will enable in a future CL.

Also cleans up remaining internal usages of deprecated APIs.

Bug: 267646347
Test: treehugger
Change-Id: Ib024eabff2a0493507523336e3d0aac8adc325f7