Copy from DisplayEventReceiver.VsyncEventData
Fix for regression.

In the regressing CL, per-frame allocation of new VsyncEventData is
removed. Instead, a single VsyncEventData is written into with new data
each frame.
However, `mLastVsyncEventData` was assigned to reference that single
VsyncEventData, so the data inside can be changed underneath. To fix,
copy the data instead of storing the reference.

Bug: 272123078
Test: ABTD run with ATP from bugs and check metric
Change-Id: I39e51cadad61b5e496effc5c7d87d824ae97438a