Use separate LogBuffer for QS config changes
There are a lot of messages logged in the QSLog LogBuffer. In some bugreports, the configChanged logs were not there anymore, as it was being overridden by others. We could have increased the size of the buffer, but using another separate one only for config changes increases the chances of having the logs we care about in bugreports.

Bug: 272012567
Test: Checked bugreport after adding this new logbuffer
Change-Id: I816b354a2530b54f418a13d0e2a309388a9146ec