Abort tf transition if window transaction is no-op
Otherwise TaskFragmentOrganizerController#onTransactionHandled
-> onTransactionFinished may trigger transition ready too early
that causes an empty transition and misses to collect the
later visibility changes, which breaks the animation.

For example, the transaction only calls createTaskFragment, but it
is failed by SecurityException for different owner task uid. And
then the client organizer will use the standard startActivity to
handle the rejection. With the abort, the failed wct can be no-op
completely and run the normal transition from startActivity.

Bug: 273422767
Test: TaskFragmentOrganizerControllerTest
Change-Id: If5e915ed3914b38273669f3d8ad715448a45f896