Isolation of Metrics for RequestSession Classes
In closing some TODOs in preparation for V4, we remove metric code from
core code as much as possbile for RequestSession. This will keep the
codebase clean and further allow metrics to be isolated for the
RequestSession classes and subclasses. In the reduction, code was
further abstracted and simplified for existing parallel TODOs.

Bug: 271135048
Test: E2E Test Conducted
Change-Id: I10aa394a365583d77118fe9ace82603b06285c44