Use IFP only for non-IME window insets providers
With the flexible insets change, it is now not possible the window is
providng insets but without a InsetsFrameProvider. The calculation of
the givenContentInsets is already included in the InsetsFrameProvider
and it is not longer necessary to calculate it in updateSourceFrame
any more.

For IME cases, the IME is provider its own frame provider to calculate
insets size with considering of the givenContentInsets. No extra
calculation needed in the general InsetsSourceProvider as well.

Tests are modified. The tests were using navigation bar to test the
mGivenContentInsets related logic. Since it will not really happen in
the production environment, modified the test to use IME window to
verify the logic, and cover the ImeSourceProvider related code.

Because the given content insets only in effect with the process of
DisplayPolicy, and is covered by the modified test is
DisplayPolicyTests, removing the other out-dated tests.

Test: DisplayPolicyTests
Test: DisplayPolicyInsetsTests
Bug: 274723551
Change-Id: I2a6828255a336fdf0d5b6c1281daff1cec02a7eb