Prevent starting split pending transition when it exists
There is a validation to check whether an empty child exists in split
from startAnimation. The IllegalStateException is thrown to say
“Somehow removed the last task in a stage outside of a proper
transition”.  There is a case to run into this situation situation
that makes a split first and launches an unsupported multi-window
activity. The onNoLongerSupportMultiWindow() is invoked and starts a
dismiss transition. Since the trampoline launch design, the
onNoLongerSupportMultiWindow() is coming and triggers another dismiss
transition. That makes the 2nd transition not consistent while the
startAnimation() is invoked.

This CL skip to set the pending transition if there is one existing.
Also apply the same protection to enter pendingTransition.

Bug: 273871464
Bug: 274835996
Test: (A|B) → C,C1 which do not support multi-window
Change-Id: I2ccff5402e9bd14cbed43a730fcd07c564725408