Introduce Communal Profile user type [minimalist]
Introduces a new user type - the Communal Profile - which is a profile
that has no parent and which is created and starts starts on boot (for
devices so configured). The profile can be accessed by any person, even
over the lockscreen and even if all human users are locked, and  is
always regarded as visible and running.

This cl does NOT include the communal profile as being part of the
profile group of other users. It is minimalist, in that it simply
creates and starts the profile; SysUi can make use of it, but external
parties (like the Launcher) would require further work to be able to do
so.

This cl is introductory. Some future tasks:
* switching users does not preserve the communal profile windows. The
  profile is still available, but each user sees its own windows.

Test: atest com.android.server.pm.UserVisibilityMediatorSUSDTest#testStartVisibleBgProfile_communalProfile
Test: atest com.android.server.pm.UserManagerTest
Test: atest UserManagerServiceUserPropertiesTest
Test: atest UserManagerServiceUserTypeTest
Test: atest UserControllerTest UserVisibilityMediatorMUPANDTest
Bug: 274838657
Bug: 276473320
Change-Id: Ib0e4e52c10f0e95913ddfde606d6995de3538f27