Change surface crop/visibilty on to-fullscreen animations
During to-fullscreen animations (either TRANSIT_EXIT_DESKTOP_MODE or
TRANSIT_CANCEL_ENTERING_DESKTOP_MODE) hide the surface control when
applying the start transaction so that it is not visible in
fullscreen size for a frame before the animation from floating size
to fullscreen starts immediately after.

Test: drag freeform task to fullscreen from desktop mode or by
cancelling a drag from fullscreen motion, verify it the task has the
right crop and doesn't show a constant fullscreen white background
as the app contents scale outwards.
Bug: 278602859

Change-Id: I51c0dfc0ca24cd8c2aa3bde1920e51f485934477