(4/N)[MediaProjection] Don't send result to server until setup ends
SysUI sends a result to MediaProjectionManagerService if the setup
flow was launched by the service, to handle projection token reuse.

If another activity is launched as part of the MediaProjection setup flow,
do not send the result to the server when the permission dialog is
dismissed; let the next activity handle sending the result.

Bug: 253218374
Test: manual
Change-Id: I92b27262a86f13c370f69fa14f427f4107c4df70