[SB Refactor] Use a real WifiRepo in MobileConnectionsRepositoryTest.
Using the real wifi repo will let us send the same network capabilities
objects to both the default network callback and the normal network
callback in tests. This makes the tests more reflective of production
scenarios.

This is in preparation for the CL that fixes b/278618530, since that CL
will rely more heavily on updating both the default and non-default
network callbacks with information.

Bug: 278618530
Test: atest MobileConnectionsRepositoryTest
Change-Id: I917506fc3c04cd63a0236cdf356cd26719c20337