Refresh activity after stronger letterboxing for camera compat
We used to refresh the activity only if the display rotation configuration was changing and the treatment was allowed. Now we are also refreshing the activity after stronger letterboxing (split screen aspect ratio is used for the camera activity). This is needed because we want to keep applying the camera compat treatment when the activity is resumed.

Fix: 277818827
Test: atest WmTests:DisplayRotationCompatPolicyTests#testOnActivityConfigurationChanging_splitScreenAspectRatioAllowed_refresh
Change-Id: I61b807f18dc89582c9e1634cd533b1fa9e364d32