Fix wallpaper flicker when quick switching apps using taskbar swipe
When quickly switching apps using the taskbar swipe questure, the
letterbox ui appears to be shown a step too late resulting in a flicker
which breifly shows the wallpaper in the background.

As this transition is now handled by shell transitions, replace 'isAnimating(TRANSITION | PARENTS)' with 'isAnimationRunningSelfOrParent()' as the former no longer detects animations run by shell.

Fix: 277887641
Test: atest WmTests:WindowStateTests
Change-Id: Ia7d18ddd3297c15dc442ea6401ac6bc14d02fa41