Removing AutomaticBrightnessStrategy refs in DPC2
This change starts the ABS/ABC move out of DPC2 by moving the
construction of ABS to BrightnessStrategySelector. Subsequent changes
will move ABC to ABS and convert the ABS calls in DPC2 into return
values of DisplayBrightnessController.updateBrightness.

This is a first step in that direction.

The call to ABS.shouldUseAutoBrightness() was changed to a value on
DisplayBrightnessState (return val of DBC.updateBrighntess()) in
order to prove out the plan of converting method calls to
return values instead. This required some additional scaffolding in
DBC which will be removed once ABS works like a standard strategy
and can fill DisplayBrightnessState during its updateBrightness phase.

Bug: 253226419
Test: atest com.android.server.display
Test: Manual, verify ScreenOffStrategy and Auto-brightness work.
Change-Id: If3251cdc76ef47effe3b8bb762d10b7311385790