Fix USB power regression when connecting USB headset during phone/voip call.
When a bi-direction USB device plugged in, first connect the output device and then the input device. The policy will prefer the paired input and output device. Connecting USB output device first can make the USB input device be used for phone/voip call once it is available.

Also cache if the USB device support jack detect to avoid opening mixer when the jack detect is not supported.

Bug: 282094846
Test: make
Test: repo steps in the bug
Change-Id: I7cc10b63c887493da284c8fad4384ee68bfd9337