Clear recents tile if exception exit case
We will try to exit split on startAnimation if stages don't have
any child tasks. But split canot handle that transition because
it will be merged to previous one.
To successful clear recents tile in this case, we need to clear
it manually before we start new dismiss transition.

Fix: 289163903
Test: manual
Test: pass existing tests
Change-Id: I1036275b9059425542b0f1a7ecff4bfd5cb0adec