Run UserLifecycleStressTest.stopManagedProfileStressTest on main user.
If switchToExistingGuestAndStartOverStressTest fails and leaves the
device with a guest user as current foreground user, it causes
stopManagedProfileStressTest -which is the next test- to fail because
it can not create a managed profile on the guest user.

This CL makes sure stopManagedProfileStressTest is run on main user,
and all test switch back to the original foreground user during the
tear down stage to leave the device in its previous state.

Bug: 290324382
Test: atest UserLifecycleStressTest
Change-Id: I694376f2f486062f4871754059dcb08ade8302e7