Move Haptic Feedback Vibration Determination Logic to Helper Class
This allows moving the logic to decide what vibration and vibration
attribute to use for a given haptic feedback type outside Window
manager and into the vibrator codespace. This is as part of the larger
effort to move haptic feedback logic out of Window manager space.

This also makes easy to add more logic to the determination of the
vibration to be used for a haptic feedback (example - allowing OEM
provided haptic effect vibration determination logic).

Bug: 185779088
Bug: 291128479
Test: manual
Change-Id: Ic00323f95ccdb77b918cc871deee19c1942359f7