Revert "Fix Rsa-Oaep operation begin on T+GSI build"
This reverts commit 6723737ede1a6bbffd8db0539878e60bd433ca5c.

Reason for revert:

Will re-introduce http://b/278157584

Even though KeyMint v2 supports the MGF_DIGEST tag, it does not include it in the key characteristics. This would not be a problem for keys generated on an Android U device with KeyMint v2 but it will be a problem on a device that was upgraded to Android U where keys were generated before the upgrade (so the MGF_DIGEST tag was not added).
Because we have no way of knowing if the MGF_DIGEST tag was specified when the key was created on KeyMint implementations older than v3, we should not add the tag on begin().

Change-Id: I1646fb72f55cf7e77b754ab48d37c8b1a6a79dc5
Merged-In: I7b34799b95eb2ff054ec4d090ccbd93e6442dcfe