Allow Communal Profile to work without HSUM.
This removes the dependency of the Communal Profile on HSUM. The
Communal Profile is no longer started in the HsumBootUserInitializer:
it now has its own initializer.

The Communal Profile will be started when user 0 is unlocked. On an HSUM
device, this occurs during boot (at around the same time that the boot
user is switched to).

Bug: 291756061
Test: build for HSUM and non-HSUM; check CP starts as expected
Change-Id: Id2ee86b0bbb67e573e0ad17e5fd57ab62cd5a13b