Fix coordinate calculation on splitscreen
The mContentRectOnScreen is a global offset i.e. the coordinate is
offset from the display origin. On the other hand, the mScreenRect
always represents a left half of the display in case of horizontally
splitted screen.
So, need to convert the mScreenRect coordinate to the offset from
the display origin.

Bug: 291004451
Test: manually done
Change-Id: I44bed2ee2988f3ce45cc49652946a2832e49ce04