Fix infinite text change flicker in shade header day/date view
Currently, we rely on onMeasure to determine the width we have available to display the day/date text. This can trigger a call to setText that leads to another layout pass, leading to an infinite loop if we receive different available widths in alternation, which is what is happening currently. Ignoring the AT_MOST measured widths prevents this loop, which should tide us over until this UI is replaced as a part of upcoming shade header modernization efforts.

Bug: 289489856
Test: manually verified behavior at different font/display size settings
Change-Id: I7a5f16a3a5d45c8face70a17651d0be4461c4c01