In HSUM configuration, allow User 0 camera access.
If a device is running in headless system user mode, then allow User 0
which is running system services and system UI camera access. This will help enable system services which needs the camera to be displayed regardless of current foreground user and not be affected by user switches.

Bug:293922840
Test: Test using automotive test application kitchensink that User 0
is able to access the camera.

Change-Id: Id0068d29320ccdf7e0724588d8eb4bbeb6de3177