Bind preview keyguard rootview to blueprint data.
Binds the preview renderer to reflect the blueprint of keyguard in its
current state. Uses a new view binder to ensure that data binding is
donw in preview renderer instead of in the sections. The data binding is
done in the renderer because arguments can be passed in the form of a
bundle.

In order to use blueprints ensure that
Flags.MIGRATE_SPLIT_KEYGUARD_BOTTOM_AREA and
Flags.MIGRATE_KEYGUARD_STATUS_VIEW are on.

Bug: 298222246
Test: adb shell cmd statusbar blueprint shortcutsBesideUdfps
Test: adb shell cmd statusbar blueprint default
Test: MIGRATE_SPLIT_KEYGUARD_BOTTOM_AREA off/on
Test: MIGRATE_KEYGUARD_STATUS_VIEW off/on

Change-Id: If6b82006adb3a25d085914107a7f5af49879b242