Fix light reveal state
An attempt to fix jank for the light reveal disconnected the internal
state from the animation start. This lead to a race condition where
the state was set to true, and expected to be set back to false after
the animation completes but in special cases the animation was
canceled. Because of this stuck state, SysUI then entered a loop where
no touches were being registered because it believed a screen off
animation in progress.

Also fix another potential race condition with the
aodUiAnimationPlaying.

Fixes: 297306022
Test: atest UnlockedScreenOffAnimationControllerTest
Test: From launcher, get phone call, hit power off
Change-Id: I754c687039b3c20fbeb1bd424d5b5b5ed0414e2f
(cherry picked from commit 94f3ec7a2c937486e8154720e6b2d544169f8fca)