Don't rebind deleted user, instead unbind all user services.
The problem is that onUserRemove we rebindService to the gone
uid that as a side effect would disable the bindings of the
current ACTIVE user(s).
This happens only if the user that we switch from is a GUEST
since GUEST users are ephemeral and will be removed onSwitch.

Test: manual
Fixes: 245254232

Change-Id: I442db90ebf8fc7b0a4bdc7d7e3236b076a6ddeb6
(cherry picked from commit cb677dd877300cee6142e35be2ecdde4e0867996)