Rename shouldUseSolidColor to allowIcon
This CL changes the name shouldUseSolidColor to allowIcon. This inverts
the meaning of the method/variable/type in ActivityRecord and WmShell,
which makes it easier to read and understand what this method does and
what the variable/type means.

Previously, the message of the shouldUseSolidColor was a "recommendation"
from WmCore to WmShell to use a solid color splash screen. This CL
transforms it into a policy for when an icon should not be used based
on the preferences provided via the public APIs.

There are no behavioural changes, only a name change.

Bug: 296354973
Test: flash && check a few splash screen interactions
Change-Id: I7a0abd040d4a551c5e4580b76a724935ee63ef1f