Handle install from unknown sources
If the source app is unknown (i.e. not holding privileged permissions to install) take a detour and ask the user to grant the app, permission to install other apps. We launch the Settings of the source app where users can toggle the permission to install apps.

When the identity of the source cannot be determined, show appropriate warning to the user before proceeding.

Bug: 182205982
Test: builds successfully
Test: No CTS Tests. Flag to use new app is turned off by default

Change-Id: I1e94e0b5c751bd0e5978120ffae04d0d215d600c