fix(window magnification): fix onSingleTap_enabled_scaleAnimates test flaky on cf fodable and tablet
The default bounce effect duration is 200ms in WindowMagnificationController. When testing onSingleTap_enabled_scaleAnimates on cuttlefish devices, sometimes the postOnAnimation may need over 200ms to dispatch and run the checking block on the mainthread, so the bounce animation has already finished and thus the scale bounce is not detected. Therefore, in WindowMagnificationControllerTest, we adjust the bounce effect duration to 2000ms, so the bounce animation check should work in expectation.

Bug: 299537784
Test: atest locally and abtd
Change-Id: Icd33debca52d1ddf5a00985e4c5d608d8370d9d1