Fix pacakge remove callback doesn't be called
In package remove case, it passes null user id list, the helper
will use mActivityManager to get running users. It looks like
cached mActivityManager is empty so the callback is ignore here.

Maybe the helper initalizes too early, try to get it and cache it
again, only skip if we cannot get it.

Bug: 302121110
Test: manual. Make sure the issue registered clients callback
be called
Test: atest PackageMonitorCallbackHelperTest
Change-Id: I4745f0436190162e8c45a09714fd016e8689aa79

Change-Id: Ib276eadf53b43e2e63c96717cee3c4b9a0a59d5c