API: add smooth switch in setFrameRateCategory
This allows the platform to have a different behavior depending on
whether a device is MRR or dVRR. When the bool is `true`, MRR devices
(those with DisplayModes that do not have vrr config) will not change
frame rates if it would cause jank. The expected usage is to mark the
bool true when an animation is running.

Bug: 300491171
Test: atest libsurfaceflinger_unittest
Change-Id: I4b28e5b100b7deecda27d4c9134e734bc9c51b47