Fix broken documentation in telephony and telecomm
Due to a bug in the build system, broken @see/@link references have not
failed to build for a while, so many errors have accumulated.

This is an attempt to patch up the telephony and telecomm docs. The
most common errors and fixes were:
- @link-ing IntDef constants from javadoc. IntDefs are not present
  in the docs, so rewrite those to natural english, or point to the
  constant prefixes
- Javadoc for public symbols referencing SystemApi symbols. Change those
  to @code instead of @link

There were various other fixes too, and I'm not sure they're all
perfect, but certainly an improvement over the brokenness we have
currently.

Bug: 303184203
Test: m docs
Change-Id: I5cefc72074ced29c0b046cf59deb7ee371e74bf0