Don't convert task windowing mode while recents is running
A task can be in running state when recents is open.
This causes issues for desktop state tracking. It relies on freeform
tasks being visible and running to determine if we are on desktop.
While recents is open, do not process handleRequest calls.
Otherwise we would process any task launches as launches on desktop and
try to convert the windowing mode accordingly.

Bug: 297590571
Test: atest DesktopTasksControllerTest

Change-Id: I5d4bd89e36980994f9877eea22c936d6d0667cfc