Fix lock workspace preview crash
The wallpaper picker app would crash when one enters the small preview,
then switches the display, for example on foldable devices, and enters
the small preview again. This is because
context.createWindowContext(...) cannot be called multiple times with
different display, or it would result in a Window Manager crash with the
message "multiple clients unsupported". In this CL, a workaround is used
to get the display size so the window context is no longer needed.

Bug: 303313708
Test: manually verified by testing the preview with different displays
Change-Id: I898fc3ca186cbefaf3cf1d653259b5acb6054ae5