Don't create new leash if there is pending position
We might want the surface position to be changed when the surface is
drawn after the buffer size is changed to prevent flicker. Creating a
new leash might update the position earlier than expected, which could
cause flicker. The CL defers updateControlForTarget to prevent creating
a new leash until the pending position is applied.

Fix: 303911439
Test: Toggle "Always show Taskbar" and make sure there is no flicker.
Change-Id: I999cb464140da382fba939711273c3fa83a223a9