Allow runtime-opaque activity to provide unspecified orientation
If the activity changes to occlude parent (opaque), the other
activities behind it will be invisible. So it is fine to respect
the requested orientation.

It still preserves the behavior that when an opaque changes to
translucent, it won't lose the ability to keep its unspecified
orientation. So the orientation won't be changed by the activity
behind during a scene transition that toggles opaque.

Fix: 305222629
Test: ActivityRecordTests#testGetOrientation,
      testShouldStartWhenMakeClientActive
Change-Id: I66b15daf70f0ef281a2cce471e52a29181bcf556