Handle uninstall result - Caller wants result back
When the caller requests for a result back, fill an intent with the uninstall status, set the correct activity result code and the activity result. The caller may also opt to get the result via UninstallCompleteCallback, in that case, send the result via the passed callback and finish the activity.

Bug: 182205982
Test: builds successfully
Test: No CTS Tests. Flag to use new app is turned off by default

Change-Id: I49a6499a57287ee22307b7e99d28882fac6444d9