Pause updating surface position while seamless rotating
Since shell transition doesn't use mPendingSeamlessRotate anymore,
it needs to check the equivalent state from AsyncRotationController
to have the same behavior as legacy transition.

If the device doesn't support HWC ScreenDecoration, it will have
non-fullscreen windows with non-zero position. Then it will hit
the case that the surface is un-rotated to previous rotation but
the position is changed to be in new rotation.

Fix: 308735235
Test: atest TransitionTests#testDisplayRotationChange
Test: Display cutout is not flickering when rotating 180 degree.

Change-Id: I6c88be283e7086d81531bc3f710e2714275129e2