Add user property for secondary auth for profiles
This change adds a user property to ensure authentication checks are
required each time a request is made to disable quiet mode for the
profile. This will be used to power the authentication checks for
private profile. This change also moves the private profile to share
credentials with parent.

Test: atest UserManagerServiceUserTypeTest. Also tested on a local build
with other private profile changes
Bug: 293571176

Change-Id: Ie1f24ae457ec05a1108608fdb531230e0f8f5acf