Remove unnecessary JNI lookups in RenderNode JNI
There was a lot of complex SDK-specific logic to initialize some
RenderNode Java accessor methods. This logic gets even more complex in
future SDK versions. However, these Java callbacks are never invoked, so
it is unnecessary to set them up. Surround that with ifdef __ANDROID__
to avoid running it on host.

Test: m -j robolectric_native_runtime dist

Change-Id: Ib81b42d0dd547f0534ee5aa7b627f487d729d09a