Refactor VisualInterruptionDecisionProviderImpl
I found a few ways to tighten up the new provider code, so... it's
refactorception.

In particular:

There are a couple bits of data we need to *log* a decision that we
don't need to *return* with it, so create a small data class to hold
that data alongside the decision itself.

Bug: 261728888
Test: atest NotificationInterruptStateProviderWrapperTest
Test: atest VisualInterruptionDecisionProviderImplTest
Flag: ACONFIG com.android.systemui.visual_interruptions_refactor DEVELOPMENT
Change-Id: Ibfd590bee4399fd5c22046bff62a23b31d17f779