Reset `mShouldClick` on ACTION_CANCEL event so click is not performed
`mShouldClick` is set to true during ACTION_DOWN event. However if
this event is interrupted or cancelled `mShouldClick` remains true so the
click is performed when the ACTION_CANCEL case is caught. Instead we
should rest `mShouldClick` to false if a motion even occurs and is
interrupted. So the click is only performed due to an ACTION_UP event.

Fix: 296921322
Test: manual
Change-Id: I9cac00ad39f1ef589eee116480d293feff4d2a11