Pass Device bus ID from native to Java
Pass Device Bus from the native side to the Java side. This is done so
we can log Device Bus ID in atoms.
Since the same vendor ID could mean different vendors depending on
whether the device is connected over bluetooth or USB, we require to know
the device bus to disambiguate the vendor IDs.

Bug: 308918242
Test: atest InputDeviceTest
Test: atest inputflinger_tests
Change-Id: I0ff2e31d7a8b610b0e1208cc5c444d1f785e162c