MediaCas : replace getService() with waitForDeclaredService()
When MediaCas service is configured to be launched as a lazy service,
getService() fails since it does not launch a service that is not
already running (which is the case with lazy services). Change the call
to waitForDeclaredService() which attempts to launch a service if not already
running.

Test: atest MediaCasTest
Bug: 301834444
Merged-In: I2c50bd8485719da4d77e1862e93129a66c102e83
Change-Id: I2c50bd8485719da4d77e1862e93129a66c102e83
(cherry picked from commit c8dd41c27adcede2543464c9d1bca5c485ae26e3)