Pointer Icon Refactor: Disable mouse position tracking in WM
The mouse position is tracked in WM so that it can provide the position
of the cursor when an app wants to update the pointer icon at will, and
when WM wants to restore the icon after the pointer leaves a window.

The refactored pointer icon pipeline provides a resolution to both these
cases, so WM will no longer need to track mouse position.

Bug: 293587049
Test: manual
Change-Id: I895e2ec4ef13cb221981ce653f977a5ed154cc9b