[flexiglass] Bouncer scene UX polish.
As per UX, aligns the top of the user switcher and the top of the
bouncer UI with each other. This is only for the tablet side-by-side layout.

Note that it still doesn't look perfectly aligned because the
placeholder user icon has built-in visual padding in the actual
drawable; but it's better than before.

Bug: 300677757
Test: before - https://screenshot.googleplex.com/3qc5qunWKA3okhA
Test: after - https://screenshot.googleplex.com/7eWytuPG6NKVUdV
Flag: NA
Change-Id: Iabe668511245a81acad3c9e2c7cf036825008c2c