Change WebView Package selection logic and error reporting
Confirmed that WebView is loaded successfully for any user who has the
default package installed and enabled even if the default package is
not installed or disabled for other users.

Some tests are changed to fit both the old and new logic. We mostly
start by user settings pointing to a not-perfect secondary package and
make sure the logic switch to the primary package.

Split testFailListingEmptyWebviewPackages to 2 smaller tests

Test:  atest com.android.server.webkit.WebViewUpdateServiceTest

Bug: 308907090

Change-Id: I5372d98d55941b59ed7b269b4f91e1fc01987f44