APIs to download and delete eSIM for admins
Updates the documenation of the download and delete APIs
to reflect the ability for enterprise admins to
manage eSIMs. Also, adds a mechanism in subscription info
to track if a sim is "owned" by an admin. This can be
used across the system to determine what actions a user
can perform on the sim (e.g. by default, users cannot
delete esims).

A separate user restriction (disallow_unmanaged_sims_globally) is
being introduced in a separate cl to block the user from
adding sims. Once merged, the permission check in
downloadSubscription can be extended to only allow the download
if the admin holds the MANAGE_DEVICE_POLICY_MOBILE_NETWORKS permission,
thus blocking download for all but admins.

Bug: 293456669
Test: btest android.devicepolicy.cts.EmbeddedSubscriptionTest
API-Coverage-Bug: 322944132
Change-Id: Id228b9ac1e1a63a519c06e7b104262ba48532373