Fix IndexOOBException when adding icons to NIC
This eliminates the indices entirely; they are unnecessary, and all
child Views will be moved to the correct index at a later time.

Specifically, the index that was used was not related to the actual
correct index of the child View; it was leftover from a former version
of the algorithm.

Flag: ACONFIG com.android.systemui.notifications_icon_container_refactor STAGING
Fixes: 317342982
Test: atest SystemUITests
Change-Id: I918bfe9c5ce4353ed164b1b96e7b3bd07bd88fa0