Enable host Windows build for librobolectric-nativeruntime
Currently SQLite JNI is disabled for Windows because both
androidfw/CursorWindow.cpp as well as the SQLite JNI registration logic
requires mmap ashmem. Supporting this is feasible but will be done in a
separate CL.

Note that some functionality still does not work 100% in Windows, such
as BitmapFactory.decodeStream and some font-related methods. These will
be resolved separately.

Test: m robolectric_native_runtime dist
Test: Robolectric ShadowNativeMatrixTest

Change-Id: Ie932ddec1298597bc3f6051574bc9dcc82b38929