Introduce a new Service#onTimeout() callback with fgsType.
This new callback will be called when an fgs type associated to a
service has exceeded its time limit. If the app does not stop the
foreground service within a short period of time after the callback,
the app will be ANR'ed.

This new time limit will be initially imposed on two types: the new
MEDIA_PROCESSING type and the DATA_SYNC type.

The previous callback with the same name will still be called for short
service, however, since that callback did not include the fgs type which
caused the timeout, this overloaded callback will be the preferred
callback going forward for all types.

Bug: 317799821
Test: builds and flashes
Test: CTS to be added
Change-Id: I63a4460c6ddaec68f75099fa1fb4d319e0b7c8e0