[flexiglass] Removes LockscreenScene.UseLockscreenContent
Since the migration from using KeyguardRootView to a compose-based
blueprint/section is almost complete, the value of keeping this constant
around is becoming less of a value and more of a burden.

Removing it is also safe because LockscreenScene is part of Flexiglass
and thus properly gated by a flag.

Bug: 316211368
Test: manually verified that System UI runs without crashing with
flexiglass on and off
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT

Change-Id: I32623f624109aae927cf85c4d9f88e452858f79d