CTA2075: Refactored the LCC API
Avoiding setting the AudioTrack. The LCC is now created by passing the
AudioAttributes and sessionId of the tracks that will receive loudness
annotated pcm values.

The new API makes it easier to send out the initial update for a media
codec by calculating the standard routing for the passed
AudioAttributes even though no AudioTrack is registered.

Test: adb shell device_config put media_audio android.media.audio.loudness_configurator_api true
Test: atest LoundessCodecHelperTest
Test: atest LoudnessCodecControllerTest
Bug: 298463873
Bug: 312529024
Bug: 314961545

Change-Id: Ic437bb6114aa3aa14debf5b443afdd0b9638ff8f