Fix capturing mid-transition frames
To stop capturing mid-transition frames, we listen to visibility changes which indicate an incoming transition to/away from the captured task, and toggle the visibility of the recorded surface accordingly. This allows the virtual display to stop sending new frames before a transition starts, so users will simply see a frozen frame from before the transition instead of the exact last frame before the window is no longer visible (which sometimes is a mid-transition frame).

Bug: 314262003
Flag: None
Test: Manual testing
Change-Id: I0a540cfa4c06b6e7cac628781c4b1773f6d6f9aa