Do not remove type string for old builds
In the change of I781bd5, multiple deprecated type string was removed.
However, the same CTS release will need to support both next QPR release
and the last release. It is not possible to support the same CTS release
with two different proto behaviors.

Add the value back with a build SDK level check, then when we release
the CTS change in the next version, the value can be removed safely.

Bug: 304398707
Bug: 319759162
Test: SplashscreenTests
Change-Id: I9711494be5ba656ac926a778dec86d6178dc663d