Merge together the two LockPatternUtilsTest
There are two LockPatternUtilsTest classes, one in coretests (module:
FrameworksCoreTests) and one in utiltests (module: FrameworksUtilTests).
It is not necessary to have two.  The "correct" place seems to be
coretests, since utiltests is intended for tests of simple Java utility
classes that aren't very closely tied to Android.  LockPatternUtils is
closely tied to Android.  Another issue with the utiltests version of
the test is that it's not currently assigned to the locksettings OWNERS.
Therefore, this CL moves the test cases of the utiltests version of the
test to the coretests version, then deletes the utiltests version.

Bug: 321737173
Ignore-AOSP-First: Conflict due to security fix ag/25939141
Test: atest FrameworksCoreTests:LockPatternUtilsTest
Change-Id: Idf562f9fa65c1d9cc0e9205217f6b8da7627f122