chore(magnification): enlarge magnification panel settings button tap area
Originally the settings button icon has margin outside of the ImageView. Tapping on the margin would be detected as tapping on underlying border view then trigger the bounce animation. Therefore, we change the outer margin to inner inset padding for the ImageView. So it will still look the same as before, and tapping on edge of the icon would become tapping on the padding, which will be detected as tapping on the button and trigger the panel as expected.

Bug: 279563444
Flag: NA
Test: manually
      atest WindowMagnificationControllerTest
Change-Id: If98eb2a12e4e714f703d56aa95aeda16333857cd