Call setFrameRate in TextureView to set frame rate properly.
Call setFrameRate in TextureView to set frame rate if one is provided.

Considering that TextureView could be invalidated with a lower
frequency comparing to the other Views. If no frame rate was set in the past 100 ms, the value of mPreferredFrameRate will be to 0 in ViewRootImpl. Otherwise, the value remains the same.

Bug: 281695725
Test: atest TextureViewTest & manual using a test app
Change-Id: Icbba5e3df8da5aa20808dd236eb93285eceddeb7