Update a11y importance together with notifStats.
The a11y importance was being updated in NSSL#addContainerView and
removeContainerView, but it uses the value from notifStats, which in
turn is updated in onAfterRenderList (which gets called AFTER
addContainerView & co.). As such, the updateImportantForAccessibility
call was using a stale value for numActiveNotifications.

Similarly, there's no point in calling
updateFooter/updateShowEmptyShadeView in addContainerView &
removeContainerView since we're also (correctly) calling them in
setNotifStats.

Fix: 318021249
Test: enable talkback and verify that there's no unlabelled view when
there are no notifs on the lockscreen
Flag: NONE

Change-Id: I1d4c4f64ee3b32c8b80c347ec30b1d588e818116