Use CanOccludePresentation state to calculate TrustedPresentationThresholds
Switch from window token to using CanOccludePresentation state to calculate
TrustedPresentationThresholds. All app windows should have this flag set. If the
window is transient, and is a system window like Toast or notification shade,
the window should not set this flag since the user or the app cannot use the
window to occlude content in a persistent manner.

Test: presubmit
Bug: b/275574214
Change-Id: Icc9b2ef19ae0c9cdc54afd643669b6154d8ad90e