Update decor insets by checking frame of all insets sources
The case is similar to [1]. But the amount of insets source window
is not changed. Such as if display cutout has a top insets with 50px.
And then status bar changes the height from 20px to 40px, but the
top insets is still 50px so nothing is updated.

By comparing the frame in each insets sources, it can still find
the differences of size, then the insets in other rotations can also
be updated.

[1]: Ieedcc6388237b7feccd8a869105a5f8e70ff4afd

This fixes flickering of display cutout when rotating the display
after increasing screen resolution. Which is caused by the additional
config change from the relayout of status bar, e.g. from above
example, there will be a 20px diff in config frame.

Fix: 319205785
Test: DisplayPolicyTests#testUpdateDisplayConfigurationByDecor
Change-Id: I6a90921d08ec6599c7cd7de27b8c1a297b0acf03