Fix privacy dot being in the wrong corner on some devices
On some devices, PrivacyDotViewController#initialize receives new view instances as parameters after a rotation event, meaning that the method continues its processing, instead of ignoring the event.

In the #initialize method, we were always using 0 as
the current rotation, when selecting a corner. This means that the top
right corner was always selected.

To fix it, we make sure to read the current rotation in #initialize.

+ Also moves PrivacyDotViewControllerTest.kt to multivalentTests/

Fixes: 321872057
Test: PrivacyDotViewControllerTest.kt
Test: Manually
Flag: NONE
Change-Id: I431d31356f7b6b354ffa5b8f86862805f40bdfb2