Generate frameworks-base-api-current.srcjar from exportable stubs
frameworks-base-api-current.srcjar is used to generate the stubs srcjar
file in dist directory during the sdk builds. Given that the sdk build
artifacts should depend on the exportable stubs instead of the
everything stubs, the src of this module should also be replaced with
the exportable stubs from the everything stubs.

Given that the only usage of this module is the stubs srcjar generation
during the sdk builds, the source of this module can be safely replaced
with the exportable stubs instead of creating a new module.

Test: patch in git_main, lunch aosp_arm-ap31-eng && m sdk dist and inspect ninja query to verify that android-stubs-src.jar does not depend on the everything stubs
Bug: 323261972
Change-Id: If37aa6ef862c69619768f91f2360b37459fd3cdb