Turn ContentProvider injection failures into errors.
Prior to this change, we simply logged this as a warning. This would
hide errors, in particular `InvocationTargetException` which itself
wraps any errors that happened during the call to `#inject`.

For the bug in question, ZenMode is actually failing to initialize,
which is causing the injection of the KeyguardSliceProvider to fail.

This change does not fix ZenMode, but will resurface the stack trace
appropriately.

Moreover, any process that is failing to initialize its
ContentProvider is already having a bad time, so if it didn't fail
here, it was going to fail somewhere else. That is what we see on the
bug.

Flag: NA
Bug: 323327880
Test: manually built and run
Change-Id: Ic2189b6c142eb0a7ec23ca49924a551626b8c326