Pass the magnetized object instead of using member variable
Wasn't able to repro this, but a couple of reports of this have come
in.

There can be a bit of time between the drag event and
dismissMagentizedObject being called (it gets called after the
bubble in the dismiss target animates out). So it could be possible
that the mMagnetizedObject has changed in that time. This CL passes
the object to the MagnetListener methods & uses that directly.

Additionally does the actions in the individual bubble / stack
listeners rather than a separate method that's only used in those
2 call sites and does something different for each one.

Flag: None; bug fix
Test: atest MagnetizedObjectTest MenuViewLayerTest DragToInteractAnimationControllerTest
Test: manual - drag and dismiss bubbles while stack is expanded
             - drag and dismiss the stack
Bug: 315127709

Change-Id: I362c824f3434e0b8296af5cb93684ed61c99378a