camera2: Move getSessionCharacteristics to CameraDeviceSetup
getSessionCharacteristics is currently a part of CameraDevice
which forces apps to call openCamera before being able to query
session characteristics. However, there is no technical reason
for this query to need an open CameraDevice.

To let apps query session characteristics without paying the
latency cost associated with openCamera, this CL moves
getSessionCharacteristics call to CameraDeviceSetup which can
be created trivially.

Bug: 320741775
Test: atest CtsCameraTestCases:CameraDeviceTest passes.
Test: atest CtsCameraTestCases:FeatureCombinationTest passes.
Test: atest CtsCameraTestCases:CameraDeviceSetupTest passes.
Change-Id: Idccea17c685c59afbdd649c8f329c7c42acc70a2