Promote setRouteVolume from SystemAPI to public API
The promotion is needed for companion apps implementing their Output
Switcher dialogs. The change is part of a wide line of work tracked in
b/192657812, which implements the necessary public APIs in MediaRouter2
for watch devices to be able to provide Output Switcher dialogs that
control media routing of the phone via their companion apps. The main
objective for the API is to provide feature parity with the phone's
Output Switcher implementation.

The child tickets of b/192657812 hold information about the other API
changes that have been made across the last few quarters towards the
same effort.

b/325323765 more context about the criticality of this API.

Bug: 325323765
Test: atest SystemMediaRouter2
API-Coverage-Bug: 325451153
Change-Id: I43b0a963d52a74e9df1779fe64baf768963ce49e