Updates FoldableDeviceStateProvider to use properties instead of flags
Updates the FoldableDeviceStateProvider to use DeviceState.Configuration
and the updated properties to describe each state. This is a refactor
of the original DeviceState flags to allow for more information
to be sent to clients to reduce the number of overlay config values
that have to be used.

Flag: N/A
Bug: 293636629
Test: FoldableDeviceStateProviderTest
Change-Id: Iec1dbd591ee1b362c7235d625a8748c022380f6e