Use the descriptor passed in when adding USB device instead of querying again.
When USB device is added, the raw descriptor has already been reported.
It is important to check if that raw descriptor is null or not to avoid
crash when creating a parser. What is more, given the descriptor has
been known when added the device, UsbHostManager can just cache that
descriptor instead of querying the raw descriptor again.

Bug: 326268581
Test: repo steps in bug
Change-Id: Iecefbd2eccbd981e77c4afe9cf8b00d830de27d3