Increasing screen brightness on home panel dream
This change manually grabs a SCREEN_BRIGHT wakelock when the home
controls dream is attached to the window, and releases it when
detached. This is a workaround due to the fact that FLAG_KEEP_SCREEN_ON
set by the DreamActivity is not honored in this case, as the home
controls activity occludes the DreamActivity.

Test: HomeControlsDreamServiceTest
Flag: ACONFIG android.service.controls.flags.home_panel_dream STAGING
Fixes: 325453103
Change-Id: I11fe981cc0ada12797248ccb58c86ac99315b707