Take #2: Shutoff calls to KeyguardStatusViewController
This will help extraneous code from running, as well as ensure
the new viewmodels are fully in control.

As part of this, make sure everything reuses the same burn-in logic
across keyguard. Also make sure that we will never adjust things
into the top offset, and log if a bad calculation is determined.

Take #2: This fixes a bad use of stateIn and optimizes use of
burnIn, to only calculate once instead of 3 times

Fixes: 327351128
Fixes: 325921793
Test: atest AodBurnInViewModelTest KeyguardIndicationAreaViewModelTest
Flag: ACONFIG com.android.systemui.migrate_clocks_to_blueprint
TEAMFOOD

Change-Id: Iac8576108c0213bca4f5b4f184690cd1219e5834