Update the oom adj seq for the service record on sending service args
When try to binds to an existing process, we may just send
the service args to them and ignored the oom adj update because
the oom adj seq for that service record isn't changed. But indeed
we should update its oom adj score because of the new connection.
So make sure we always update the oom adj seq on sending service args.

Bug: 332435378
Test: Manual - Verify the keyboard haptics
Change-Id: I80af8095fa9003a96339ec4c4843506c704d9d87