Use "C" numeric locale for consistency in strtof
In Android, strtof is used in places such as PathParser to parse floats
in path expressions in drawables, i.e. "C0.2,0 0.1,1 0.5, 1 L 1,1".

In certain locales (e.g. Germany), commas are interpreted as part of
numbers, so this is not parsed correctly.

Use the minimal "C" numeric locale to ensure consistency when doing
strtof and other floating point parsing operations.

This is also done in LayoutlibLoader.

Bug: 334093536
Test: ShadowNativePathParserTest in Google3
Change-Id: I21e7115240045ebf065362a6e28f6af710509a13