Heads up cycling animation
Introduce a new type of hun intro and outro animation - the hun
cycling animation. This animation is for the transition from the old
hun to the new hun when the avalanche notification feature is enabled,
and at least two huns are posted within a short period of time.

Bug: 316404716
Test: manual
Flag: ACONFIG notification_heads_up_cycling DEVELOPMENT
Change-Id: I9d1596075af2bef2b50e6b5ff07503647ee1b0ba