Notify keyguardGoingAway prior to making the call.
This is the ordering used in the keyguardGoingAwayRunnable, but we had it swapped for showSurfaceBehindKeyguard (which is used during swipe to unlock). ag/25832421 enforced that prepareForUnlock() must be called prior to Launcher's onStart(), and onStart() is triggered by keyguardGoingAway. This resulted in prepareForUnlock() having no effect, and the animation not running.

This should be safe since all other unlock flows already use a code path that notifies goingAway=true prior to the actual call.

Fixes: 334046148
Test: swipe to unlock
Flag: NA
Change-Id: Iaffdd6a45db5c03dc3f137b3dab6e209d3a57c52
(cherry picked from commit 93f49b79c890e0a7716c62af65437e23c6563de3)