Restore communal widgets after host is restored
This change adds an observer to the app widget host restored broadcast.
When this signal is received, a backed-up state is expected to have been
written on disk and read into memory. The widget ids are updated based
on the mapping provided by the broadcast, and database is wiped and
restored to the backed up state.

Test: atest CommunalWidgetDaoTest
Test: atest CommunalWidgetRepositoryImplTest
Test: atest CommunalBackupRestoreStartableTest
Test: manual; see instructions at go/glanceable-hub-br
Bug: 309809222
Flag: ACONFIG com.android.systemui.communal_hub TEAMFOOD
Change-Id: Ia67bb901e6976d1b7507672e58b53b6bf5b42c81