Optimise UserWakeupStorePerformance
Before the change, mStartingUsers list was used to store users that were being started. This caused spikiness and user boot time regression.

After the change, this intermediate mStartingUsers list was removed and the performance was returned to previous state. This step is not necessary as it lasts only a few ms which leaves a very narrow window when user ID is not in the list.

Bug: 337788578
Test: manual && perfetto trace testing && atest UserWakeupStoreTest
Flag: com.android.server.alarm.start_user_before_scheduled_alarms
Change-Id: I83ad157c0b0fa118ce9f2bd3237a44e784641132