Add action revealing to screenshot entrance animation.
The actions vertical entrance animation is a single animation, but
at a 200ms milestone during this animation, additional actions may be
(horizontally) revealed.

This change does two things:
1. Give each action a showDuringEntrance bool that indicates whether it
   should be included in the container from the very beginning.
2. 200ms into the actions entrance, add the remaining actions to the
   actions bar.

This puts some animation state into the viewmodel, which really should
be internal to the viewbinder, but we're not ready to move that due to
various flag/legacy constraints.

Bug: 329659738
Test: Manual inspection with slowed-down animations
Flag: ACONFIG com.android.sysui.screenshot_shelf_ui2 TRUNKFOOD
Change-Id: Ide4ee39117a5e80eaaa981a3ca10f40924c49ef9