Add flag to match header density with the desktop windows'
In desktop windowing, tasks by default have a custom density applied to
them to optimize for the available display space. This change updates
the App Header's density to also use that custom density instead of
SysUI's density. This allows the header height to stay consistent on
density changes and allows apps using custom headers to more easily draw
content in the header.

Bug: 332414819
Test: m
Change-Id: I8345e6e530bd57b60f7374e5d08d7e307b0efc26