Fix wildcard scene edges
When scene container is flipped wildcard edges with UNDEFINED are not
allowed because KTF will not track all edges in or out of e.g.
Scenes.Gone but rather only the edges between Lockscreen and Gone.

This CL fixes some occurrences but not yet all.

Test: None
Bug: b/330311871
Flag: com.android.systemui.scene_container
Change-Id: Ia09d95fdf15768b24d5e6c2103697574b9458d47