Don't handle interruptions too aggressively
This CL simplifies Element.prepareInterruption() so that we simply
reconcile between SceneStates that are part of the same transition (the
new or interrupted one). This means that we animate transition jumps
less than before, but I think that this is overall what we want.

See b/290930950#comment28 for details and before/after videos.

Bug: 290930950
Test: ElementTest. Existing tests already cover the use cases we are
 interested in.
Flag: com.android.systemui.scene_container
Change-Id: I096625fd1e05a470b849ad046e28b2284adb1228