Fix various hwui compile errors when building with Android toolchain
Attempting to run `m libhwui` causes compile errors in the
android12-hostruntime-dev repo. This is because hwui has been modified
primarily to enhance host graphics, but as a result trying to build it
using the Android toolchain causes compile errors.

Update libhwui to fix various compile errors so it's possible to run `m
libhwui`.

Test: m libhwui
Bug: 344979295
Change-Id: I74a7d09e2f7dd8249c8e6e692e6783cfb70bc6f8