Support partial viewer config loading and unloading
Most of the time we only enable some groups to be logged to protolog and and not all group. So instead of loading the entire config in memory as soon as any group is being logged to protolog we support more granular control for only loading target groups.

Flag: android.tracing.perfetto_protolog_tracing
Bug: 352290057
Test: atest com.android.internal.protolog.ProtoLogViewerConfigReaderTest
Change-Id: I6d3fb041c5d3947a59a1de4c39660fe65a3a0051