Validate Event Assigner for input devices other than touchscreen
Extended the logic for processing MOVE events that happen after DOWN events in the same frame to all types of Motion events and other input devices.
If a sequence like DOWN -> MOVE happens in a single frame we want to provide the eventId of the DOWN event.
Before this CL, this was implemented only for Move events coming from Touchscreen. This CL implements this behaviour for all Motion Events and other input devices such as Stylus, Mouse, Touchpad

Added tests for this functionality and also to test that if a DOWN -> EVENT -> UP/CANCEL happens the DOWN event is processed.

Bug: b/234026254
Test: atest InputTests
Flag: EXEMPT bugfix
Change-Id: I50b6035bd5de44845ef88484e148120b201a149e