Adding a SliderDragVelocityProvider for slider haptics
The provider is a small abstraction around different methods to obtain
the velocity with which a slider is being dragged. This can make the
SliderHapticFeedbackProvider usable for SeekBars as well as Slider
composables. Velocity computation can be done in any arbitrary unit
instead of only pixels. This is now reflected on the
SliderHapticFeedbackConfig docstring

Test: atest SystemUITests:SliderHapticFeedbackProviderTest
Flag: NONE minor refactor to adapt for future usages
Bug: 356389497
Change-Id: I4d31923c2cf9eecb138dac36af02d019c9a6f831