Handle autocork message on the background thread
The autocork feature implements a timeout with messages on a
looper.  When the timeout expires, the message handler sets a system
property.  This is I/O (socket read/write).

The original design handled these messages on the main thread but the
I/O causes delays.  This change moves message handling to the
background thread, which is less sensitive to delays.

Flag: EXEMPT bugfix
Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests
Bug: 356335509
Change-Id: Ia8a416726f6aa6e9067c92ef4fd998211fc66eae