Work around not working "ime set <imeId>"
This CL adds a debug log and workaround for the case when

  adb shell ime set --user <userId> <imeId>

does not update "default_input_method" secure settings in quite low
probability.

Note that the above command is intended only for developers and
testing purposes. There must be no user observable behavior change in
this CL.

Bug: 354782333
Test: presubmit
Flag: TEST_ONLY
Change-Id: Ia14a43f901c7a12c4a51cd747a71bde61e3157bb