[flexiglass] Attempt device entry when told to dismiss by system server
When system server asks System UI to dismiss the keyguard, attempt
device entry.

Before this change, there would be a crash from unexpectedly calling
legacy bouncer code while on Flexiglass.

With this change, the bouncer or alternate bouncer show up as needed.

The bug described in the attached bug is still occuring however: back
navigation on the alternate bouncer affects the scene containe and not
the alternate bouncer. Working on that next.

Bug: 360109685
Test: manually verified that the bouncer shows up if fingerprint isn't
set up, the alternate bouncer shows up if fingerprint is set up
Flag: com.android.systemui.scene_container

Change-Id: I3989a3df475f1245ece4d47c6516f7140dfc9d39