Flexiglass: use Interactors for remote input callOnClick logic
When the user clicks a remote input while the device is locked,
StatusBarRemoteInputCallback is responsible for displaying the bouncer,
waiting for the user to unlock the device, and then "re-clicking" the
remote input to finish whatever task the user started.

When Flexi is enabled, use TransitionState and
DeviceUnlockedStatus.isUnlocked to determine when the device unlock
process has reached a point where it is safe to "re-click" the remote
input.

Bug: 359530769
Test: presubmit
Flag: com.android.systemui.scene_container
Change-Id: I370534ea0710fa947666f97e363c053211ab6ac1