Replaced verifyZeroInteractions with verifyNoMoreInteractions
The on-device unit tests are currently using Mockito 2.23, and the Robolectric tests use mockito 4.8.1. In 2.23, verifyZeroInteractions had already been deprecated, and in 4.8.1, it was removed entirely.  In order to move the tests to multivalentTests, we had to blanket-change the API, and there were some files whose runners hadn't been changed from AndroidTestingRunner to AndroidJUnit4.

MagnificationModeSwitchTest.java was mistakenly moved to the multivalentTests directory, where it was failing at runtime under Robolectric (b/369575544). This CL moves it back to the tests directory, and other files had to be moved back in ag/29583402.

Test: tested manually with atest
Bug: 369888148
Flag: TEST_ONLY
Change-Id: I04f9494d914b3e7627c2e773e49cdde781935880