Fix back gesture showing up on hub on lock screen
The notification shade is always considered expanded and visible while
on the lock screen, causing the back gesture to show up on the hub.

This CL adjusts the criteria to look for quick settings being expanded,
which is always true on tablet when the shade is opened in landscape.

This does mean that back gesture will not work when opening the shade
over the hub in portrait mode, filed 370108274 to track.

Bug: 370108262
Fixed: 370108262
Test: atest QuickStepContractTest
      also manually verified on device
Flag: com.android.systemui.communal_hub
Change-Id: I5762166e51b49e8233f6e9cc7f085bdf732d1d9e