Create ManagerRecord#notifyRoutesUpdated
This change is part of a series of non-functional refactors to
encapsulate binder operations in the record that corresponds to the
target binder object. The objective is to improve the logging associated
with binder failures by adding information that's available in the
record. For example: Package names and ids.

Bug: 360129098
Test: atest CtsMediaBetterTogetherTestCases MediaRouter2HostSideTest
Flag: EXEMPT refactor
Change-Id: Iba0897590d0129ffb0eaf8e152713d582be1fd1f