Cancel multipress gesture if a press is too slow
If a user starts a multipress gesture (i.e. starts a second/third press)
of a button, but that second/third press is long, cancel the multipress
gesture. We require the first and second (and third) down events to
happen quickly after one another, so it is more organic that we expect
the consecutive presses to be within the same duration.

Bug: 370095426
Test: atest SingleKeyGestureTests
Flag: com.android.hardware.input.abort_slow_multi_press

Change-Id: Id17c13e7d7863e3f7949da7b53ee7f4dfc0c1236