Set Launcher alpha to 0 during Desktop exit transition.
During the Desktop last-window-close exit transition we are
simultaneously
1. closing a Desktop app
2. closing DesktopWallpaperActivity
3. opening LauncherActivity

Since the Desktop app might not cover the entire screen, and
DesktopWallpaperActivity is transparent - the user will see
LauncherActivity beneath those tasks/windows throughout the exit
animation.

Before this CL the alpha of the leash of LauncherActivity was set to 1
at the beginning of the animation (before the Launcher Workspace view
had time to reset its Views), causing a flicker showing the Workspace
view (icons and widgets) for a single frame.
With this CL we set the alpha of the Launcher leash to 0 at the start of
the Desktop exit animation, so the LauncherActivity and its Workspace
won't show until the Workspace view is ready.

Test: manual
Bug: 366153364
Flag: com.android.window.flags.enable_desktop_windowing_exit_transitions
Change-Id: I8e4d7ef01052e2158a644630f85d2f9ec4e9bda4