Add mixed transition handling for desktop changes
Identifies desktop changes (immersive exits for now) within transitions
that should be mainly handled by existing handlers and splits it so that
the desktop change is animated by its own handler.

Specifically, deskop immersive exits during an open/front transition
should be split so that the immersive task that is being resized is
animated by the immersive handler, while the opening task remains
animated by the DefaultTransitionHandler.
Similarly, for intent launches with a remote handler, the immersive
change is split and animated by the immersive handler and the rest is
sent to the remote handler to animate.

Flag: com.android.window.flags.enable_fully_immersive_in_desktop
Bug: 372319492
Test: while in desktop immersive, launch an app from the taskbar and see
both of them animate
Test: while in desktop immersive, launch an app from the notification
shade and see both of them animate

Change-Id: Ib63601322df20505dd7c012d85b5705d31dec5a8